+++ packages/mvc/src/control.ts
@@ -1,3 +1,5 @@
+import type { Model } from "./model";
+
 /**
  * Update callback function.
  * 
@@ -23,7 +25,7 @@ declare namespace Observable {
 }
 
 interface Observable {
-  [Observable](callback: Observable.Callback, required?: boolean): this;
+  [Observable](callback: Observable.Callback, required?: boolean): this | void;
 }
 
 const Observable = Symbol("observe");
@@ -149,9 +151,9 @@ function enqueue(eventHandler: (() => void)){
  * @param callback - Function to invoke when values change.
  * @param requireValues - If `true` will throw if accessing a value which is `undefined`.
  */
-function createEffect<T extends Observable>(target: T, callback: Effect<Required<T>>, requireValues: true): () => void;
-function createEffect<T extends Observable>(target: T, callback: Effect<T>, recursive?: boolean): () => void;
-function createEffect<T extends Observable>(target: T, callback: Effect<T>, argument?: boolean){
+function createEffect<T extends Model>(target: T, callback: Effect<Required<T>>, requireValues: true): () => void;
+function createEffect<T extends Model>(target: T, callback: Effect<T>, recursive?: boolean): () => void;
+function createEffect<T extends Model>(target: T, callback: Effect<T>, argument?: boolean){
   let unset: ((update: boolean | null) => void) | undefined;
   let reset: (() => void) | null | undefined;