+++ packages/react/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.69.2](https://github.com/gabeklein/expressive-mvc/compare/v0.69.1...v0.69.2) (2025-09-11)
+
+
+### Bug Fixes
+
+* update Props type to omit Model keys for better compatibility ([77a1cbe](https://github.com/gabeklein/expressive-mvc/commit/77a1cbef3faddeebab5f57e034ad798556a8325f))
+
+
+
+
+
+## [0.69.1](https://github.com/gabeklein/expressive-mvc/compare/v0.69.0...v0.69.1) (2025-09-10)
+
+**Note:** Version bump only for package @expressive/react
+
+
+
+
+
+# [0.69.0](https://github.com/gabeklein/expressive-mvc/compare/v0.68.0...v0.69.0) (2025-09-05)
+
+
+### Bug Fixes
+
+* support assigning function components as render method ([fade81e](https://github.com/gabeklein/expressive-mvc/commit/fade81eba5a049cc1aa30366823516142bc449dc))
+
+
+
+
+
+# [0.68.0](https://github.com/gabeklein/expressive-mvc/compare/v0.67.0...v0.68.0) (2025-08-21)
+
+
+### Features
+
+* **react:** add suspense fallback to Provider ([c334366](https://github.com/gabeklein/expressive-mvc/commit/c33436628c0bcdcec1c4b161feca225c8c248dd0))
+
+
+
+
+
 # [0.67.0](https://github.com/gabeklein/expressive-mvc/compare/v0.66.2...v0.67.0) (2025-07-28)
 
 **Note:** Version bump only for package @expressive/react